[AC-5495] - refactor 'make help' information to reflect the             addition of the TAG argument           - move 'make release' logic into the 'create_release.sh'           - add conditional that creates a semantic tag if a custom             one is not provided, otherwise it uses the custom one             and tags all the repos.

+++ create_release.sh
@@ -1,7 +1,20 @@
-docker run --rm  -v"$(pwd)":/app  -ti semantic-release  -- semantic-release version
-export pwd=$(pwd)
-export TAG=$(docker run --tty=false --rm  -v$pwd:/app  -i semantic-release -- semantic-release version --noop | grep "Current version: " | cut -d ' ' -f 3 | sed -e "s/\r//")
+
+# if a tag hasn't been passed in as a parameter, create a semantic one
+if [ ! -n "$TAG" ]; then
+    git commit --allow-empty -m "generating a new release"
+	git push
+    docker run --rm  -v"$(pwd)":/app  -ti semantic-release  -- semantic-release version
+    export pwd=$(pwd)
+    export TAG=$(docker run --tty=false --rm  -v$pwd:/app  -i semantic-release -- semantic-release version --noop | grep "Current version: " | cut -d ' ' -f 3 | sed -e "s/\r//")
+else
+    echo "You passed in '$TAG' as the custom tag"
+    # TR == test release
+    echo "We're going to refactor the tag into 'TR-$TAG' to avoid any conflicts."
+    TAG="TR-$TAG"
+fi
+
 echo $TAG
+
 git push
 git push --tags
 cd ../django-accelerator && git tag "v${TAG}"
